+++ packages/shared-src/src/reports/index.js
@@ -1,7 +1,7 @@
 import * as admissions from './admissions';
 import * as incompleteReferrals from './incomplete-referrals';
 import * as recentDiagnoses from './recent-diagnoses';
-import * as covidVaccineList from './covid-vaccine-list';
+import * as vaccineList from './vaccine-list';
 import {
   generateCovidVaccineSummaryDose1Report,
   generateCovidVaccineSummaryDose2Report,
@@ -24,8 +24,8 @@ export function getReportModule(reportType) {
       return incompleteReferrals;
     case 'recent-diagnoses':
       return recentDiagnoses;
-    case 'covid-vaccine-list':
-      return covidVaccineList;
+    case 'vaccine-list':
+      return vaccineList;
     case 'covid-vaccine-summary-dose1':
       return {
         permission: covidVaccineSummaryPermission,
@@ -50,3 +50,5 @@ export function getReportModule(reportType) {
       return covidSwabLabTestsSummary;
   }
 }
+
+export { REPORT_DEFINITIONS } from './reportDefinitions';