+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 ### Features ⭐
 
 - Add registered patients line-list report
+- Generic Vaccine line list report
+- Allow disabling reports through server config
 
 ## Tweaks ⚖️